Move ViewModels under viewmodel/
