+++ src/main/java/com/atlauncher/gui/tabs/settings/CommandsSettingsTab.java
@@ -27,6 +27,8 @@ import javax.swing.JTextField;
 import javax.swing.JTextPane;
 
 import com.atlauncher.listener.DelayedSavingKeyListener;
+import com.atlauncher.viewmodel.base.settings.ICommandsSettingsViewModel;
+import com.atlauncher.viewmodel.impl.settings.CommandsSettingsViewModel;
 import org.mini2Dx.gettext.GetText;
 
 import com.atlauncher.App;
