+++ src/main/java/com/atlauncher/viewmodel/base/settings/IGeneralSettingsViewModel.java
@@ -15,8 +15,9 @@
  * You should have received a copy of the GNU General Public License
  * along with this program. If not, see <http://www.gnu.org/licenses/>.
  */
-package com.atlauncher.gui.tabs.settings;
+package com.atlauncher.viewmodel.base.settings;
 
+import com.atlauncher.gui.tabs.settings.GeneralSettingsTab;
 import com.atlauncher.utils.sort.InstanceSortingStrategies;
 
 import java.util.List;
@@ -249,16 +250,16 @@ public interface IGeneralSettingsViewModel extends IAbstractSettingsViewModel {
     /**
      * Data object for UI representation.
      */
-    class LauncherTheme {
+    public class LauncherTheme {
         /**
          * Identification of the theme
          */
-        final String id;
+        public final String id;
 
         /**
          * UI Label for the theme
          */
-        final String label;
+        public final String label;
 
         public LauncherTheme(String id, String label) {
             this.id = id;
