+++ src/main/java/com/atlauncher/viewmodel/base/settings/IJavaSettingsViewModel.java
@@ -15,10 +15,11 @@
  * You should have received a copy of the GNU General Public License
  * along with this program. If not, see <http://www.gnu.org/licenses/>.
  */
-package com.atlauncher.gui.tabs.settings;
+package com.atlauncher.viewmodel.base.settings;
 
 import com.atlauncher.constants.Constants.ScreenResolution;
 import com.atlauncher.data.CheckState;
+import com.atlauncher.gui.tabs.settings.JavaSettingsTab;
 import org.jetbrains.annotations.Nullable;
 
 import java.util.List;
