+++ src/main/java/com/atlauncher/viewmodel/base/settings/IModsSettingsViewModel.java
@@ -15,11 +15,12 @@
  * You should have received a copy of the GNU General Public License
  * along with this program. If not, see <http://www.gnu.org/licenses/>.
  */
-package com.atlauncher.gui.tabs.settings;
+package com.atlauncher.viewmodel.base.settings;
 
 import com.atlauncher.data.AddModRestriction;
 import com.atlauncher.data.InstanceExportFormat;
 import com.atlauncher.data.ModPlatform;
+import com.atlauncher.gui.tabs.settings.ModsSettingsTab;
 
 import java.util.function.Consumer;
