+++ src/main/java/com/atlauncher/viewmodel/impl/settings/CommandsSettingsViewModel.java
@@ -15,11 +15,12 @@
  * You should have received a copy of the GNU General Public License
  * along with this program. If not, see <http://www.gnu.org/licenses/>.
  */
-package com.atlauncher.gui.tabs.settings;
+package com.atlauncher.viewmodel.impl.settings;
 
 import com.atlauncher.App;
 import com.atlauncher.evnt.manager.SettingsManager;
 import com.atlauncher.evnt.manager.SettingsValidityManager;
+import com.atlauncher.viewmodel.base.settings.ICommandsSettingsViewModel;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
