+++ src/main/java/com/atlauncher/viewmodel/impl/settings/LoggingSettingsViewModel.java
@@ -15,10 +15,11 @@
  * You should have received a copy of the GNU General Public License
  * along with this program. If not, see <http://www.gnu.org/licenses/>.
  */
-package com.atlauncher.gui.tabs.settings;
+package com.atlauncher.viewmodel.impl.settings;
 
 import com.atlauncher.App;
 import com.atlauncher.evnt.manager.SettingsManager;
+import com.atlauncher.viewmodel.base.settings.ILoggingSettingsViewModel;
 
 import java.util.function.Consumer;
