+++ src/main/java/com/atlauncher/viewmodel/impl/settings/ModsSettingsViewModel.java
@@ -15,13 +15,14 @@
  * You should have received a copy of the GNU General Public License
  * along with this program. If not, see <http://www.gnu.org/licenses/>.
  */
-package com.atlauncher.gui.tabs.settings;
+package com.atlauncher.viewmodel.impl.settings;
 
 import com.atlauncher.App;
 import com.atlauncher.data.AddModRestriction;
 import com.atlauncher.data.InstanceExportFormat;
 import com.atlauncher.data.ModPlatform;
 import com.atlauncher.evnt.manager.SettingsManager;
+import com.atlauncher.viewmodel.base.settings.IModsSettingsViewModel;
 
 import java.util.function.Consumer;
