+++ src/main/java/com/atlauncher/viewmodel/impl/settings/NetworkSettingsViewModel.java
@@ -15,21 +15,14 @@
  * You should have received a copy of the GNU General Public License
  * along with this program. If not, see <http://www.gnu.org/licenses/>.
  */
-package com.atlauncher.gui.tabs.settings;
+package com.atlauncher.viewmodel.impl.settings;
 
 import com.atlauncher.App;
 import com.atlauncher.Network;
 import com.atlauncher.evnt.manager.SettingsManager;
-import com.atlauncher.data.CheckState;
 import com.atlauncher.evnt.manager.SettingsValidityManager;
-import com.atlauncher.utils.Utils;
-import org.apache.logging.log4j.LogManager;
-import org.apache.logging.log4j.Logger;
+import com.atlauncher.viewmodel.base.settings.INetworkSettingsViewModel;
 
-import java.net.InetSocketAddress;
-import java.net.Proxy;
-import java.util.Set;
-import java.util.concurrent.TimeUnit;
 import java.util.function.Consumer;
 
 /**
